refactor: Correction of identification and change of names of positions and variables
