+++ src/components/CardPoster/card-poster.vue
@@ -3,8 +3,8 @@ import { Streamings } from '@/models/streaming.model';
 
 export default {
     name: "Card-Poster",
-    props:{
-        url: Streamings,
+    props: {
+        streamingPosters: Streamings,
     }
 }
 </script>
@@ -12,7 +12,7 @@ export default {
 <template>
     <Card class="w-[304px] h-[456px] overflow-hidden rounded-2xl">
         <template #header>
-            <img alt="user header" :src="`https://image.tmdb.org/t/p/w342/${url?.poster_path}`"/>
+            <img alt="poster header" :src="`https://image.tmdb.org/t/p/w342/${streamingPosters?.poster_path}`"/>
         </template>
     </Card>
 </template>
