+++ src/components/NavBar/nav-bar.vue
@@ -24,8 +24,8 @@ export default {
             </div>
             <div class="block sm:hidden">
                 <Button @click="toggle" class="flex items-center px-3 py-2 border rounded hover:text-white hover:border-white">
-                        <v-icon class="fill-current h-3 w-3" viewBox="0 0 20 20" name="hi-menu"/>
-                    </Button>
+                    <v-icon class="fill-current h-3 w-3" viewBox="0 0 20 20" name="hi-menu"/>
+                </Button>
             </div>
             <div :class="menu ? 'block' : 'hidden'" class="w-full flex-grow sm:flex sm:items-center sm:w-auto">
                 <div class="text-sm sm:flex-grow">
@@ -42,9 +42,9 @@ export default {
                 <div>
                     <RouterLink class="no-underline inline-block text-base px-4 py-2 leading-none border rounded text-white border-white hover:border-transparent hover:bg-white hover:text-red-700 mt-4 sm:mt-0" to="/favorite">
                         <span>
-                                Favoritos
-                                <v-icon name="ri-heart-fill" />
-                            </span>
+                            Favoritos
+                            <v-icon name="ri-heart-fill" />
+                        </span>
                     </RouterLink>
                 </div>
             </div>
