+++ src/components/StructurePages/structure-pages.vue
@@ -1,10 +1,10 @@
 <script lang="ts">
-import type { StreamingService } from '@/services/streaming.service';
+import type { StreamingService } from '@/views/streaming.service';
 
 export default {
     name: "StructurePages",
     props: {
-        url: Array < StreamingService > ,
+        streamingPosters: Array < StreamingService > ,
     }
 }
 </script>
@@ -13,11 +13,11 @@ export default {
     <div class="min-h-screen bg-gray-800">
         <div class="flex items-center text-center sm:pt-8">
             <ul class="flex-wrap">
-                <li class="inline-block relative lg:m-6 lg:mt-4 sm:justify-center my-5" v-for="posters in url">
+                <li class="inline-block relative lg:m-6 lg:mt-4 sm:justify-center my-5" v-for="posters in streamingPosters">
                     <div>
                         <div class="transform transition duration-500 hover:scale-110 ">
                             <RouterLink to="/details">
-                                <card-poster class="mx-5" :url="posters"/>
+                                <card-poster class="mx-5" :streamingPosters="posters"/>
                             </RouterLink>
                         </div>
                     </div>
